Split up `Router` in preparation for adding images
