+++ api/api/src/main/scala/uk/ac/wellcome/platform/api/Router.scala
@@ -1,9 +1,7 @@
 package uk.ac.wellcome.platform.api
 
-import scala.concurrent.{ExecutionContext, Future}
-import scala.util.{Failure, Success}
+import scala.concurrent.ExecutionContext
 import akka.http.scaladsl.model.{HttpEntity, MediaTypes}
-import akka.http.scaladsl.model.StatusCodes._
 import akka.http.scaladsl.server.{
   Directives,
   MalformedQueryParamRejection,
@@ -11,35 +9,17 @@ import akka.http.scaladsl.server.{
   Route,
   ValidationRejection
 }
-import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport
-import com.sksamuel.elastic4s.{ElasticClient, ElasticError, Index}
-import io.circe.Printer
-import grizzled.slf4j.Logger
-
-import uk.ac.wellcome.platform.api.services.{
-  CollectionService,
-  ElasticsearchService,
-  WorksService
-}
+import com.sksamuel.elastic4s.ElasticClient
 import uk.ac.wellcome.elasticsearch.ElasticConfig
-import uk.ac.wellcome.platform.api.elasticsearch.ElasticsearchErrorHandler
 import uk.ac.wellcome.platform.api.swagger.SwaggerDocs
-import uk.ac.wellcome.models.work.internal._
 import uk.ac.wellcome.platform.api.models._
-import uk.ac.wellcome.display.models._
-import uk.ac.wellcome.display.models.v2._
-import uk.ac.wellcome.display.models.Implicits._
-import uk.ac.wellcome.display.json.DisplayJsonUtil
+import uk.ac.wellcome.platform.api.rest.{CustomDirectives, WorksController}
 
 class Router(elasticClient: ElasticClient,
              elasticConfig: ElasticConfig,
-             apiConfig: ApiConfig)(implicit ec: ExecutionContext)
-    extends FailFastCirceSupport
-    with Directives
-    with Tracing {
-
-  import MultipleWorksResponse.encoder
-  import ResultResponse.encoder
+             implicit val apiConfig: ApiConfig)(implicit ec: ExecutionContext)
+    extends Directives
+    with CustomDirectives {
 
   def routes: Route = handleRejections(rejectionHandler) {
     ignoreTrailingSlash {
@@ -48,14 +28,10 @@ class Router(elasticClient: ElasticClient,
           pathPrefix("v2") {
             concat(
               path("works") {
-                MultipleWorksParams.parse { params =>
-                  getWithFuture(multipleWorks(params))
-                }
+                MultipleWorksParams.parse { worksController.multipleWorks }
               },
-              path("works" / Segment) { workId =>
-                SingleWorkParams.parse { params =>
-                  getWithFuture(singleWork(workId, params))
-                }
+              path("works" / Segment) { workId: String =>
+                SingleWorkParams.parse { worksController.singleWork(workId, _) }
               },
               path("context.json") {
                 getFromResource("context-v2.json")
@@ -66,7 +42,12 @@ class Router(elasticClient: ElasticClient,
             )
           },
           path("v1" / Remaining) { _ =>
-            v1ApiGone
+            gone(
+              """"
+                |This API is now decommissioned.
+                |Please use https://api.wellcomecollection.org/catalogue/v2/works.
+              """.stripMargin.replace('\n', ' ')
+            )
           },
           pathPrefix("management") {
             concat(
@@ -85,126 +66,16 @@ class Router(elasticClient: ElasticClient,
     }
   }
 
-  def multipleWorks(params: MultipleWorksParams): Future[Route] =
-    transactFuture("GET /works") {
-      val searchOptions = params.searchOptions(apiConfig)
-      val index = params._index.map(Index(_)).getOrElse(elasticConfig.index)
-      worksService
-        .listOrSearchWorks(index, searchOptions)
-        .map {
-          case Left(err) => elasticError(err)
-          case Right(resultList) =>
-            extractPublicUri { uri =>
-              complete(
-                MultipleWorksResponse(
-                  resultList,
-                  searchOptions,
-                  params.include.getOrElse(V2WorksIncludes()),
-                  uri,
-                  contextUri
-                )
-              )
-            }
-        }
-    }
+  lazy val worksController =
+    new WorksController(elasticClient, apiConfig, elasticConfig)
 
-  def singleWork(id: String, params: SingleWorkParams): Future[Route] =
-    transactFuture("GET /works/{workId}") {
-      val index = params._index.map(Index(_)).getOrElse(elasticConfig.index)
-      val includes = params.include.getOrElse(V2WorksIncludes())
-      worksService
-        .findWorkById(id)(index)
-        .flatMap {
-          case Right(Some(work: IdentifiedWork)) =>
-            if (includes.collection) {
-              val expandedPaths = params._expandPaths.getOrElse(Nil)
-              retrieveTree(index, work, expandedPaths).map {
-                workFound(work, _, includes)
-              }
-            } else
-              Future.successful(workFound(work, None, includes))
-          case Right(Some(work: IdentifiedRedirectedWork)) =>
-            Future.successful(workRedirect(work))
-          case Right(Some(_)) => Future.successful(workGone)
-          case Right(None)    => Future.successful(workNotFound(id))
-          case Left(err)      => Future.successful(elasticError(err))
-        }
-    }
-
-  def workFound(work: IdentifiedWork,
-                tree: Option[(Collection, List[String])],
-                includes: V2WorksIncludes): Route =
+  def swagger: Route = get {
     complete(
-      ResultResponse(
-        context = contextUri,
-        result = DisplayWorkV2(work, includes).copy(
-          collection = tree.map {
-            case (tree, expandedPaths) =>
-              DisplayCollection(tree, expandedPaths)
-          }
-        )
-      )
-    )
-
-  def workRedirect(work: IdentifiedRedirectedWork): Route =
-    extractPublicUri { uri =>
-      val newPath = (work.redirect.canonicalId :: uri.path.reverse.tail).reverse
-      redirect(uri.withPath(newPath), Found)
-    }
-
-  def workGone: Route = error(
-    DisplayError(
-      ErrorVariant.http410,
-      Some("This work has been deleted")
-    )
-  )
-
-  def workNotFound(id: String): Route = error(
-    DisplayError(
-      ErrorVariant.http404,
-      Some(s"Work not found for identifier ${id}")
-    )
-  )
-
-  def v1ApiGone = error(
-    DisplayError(
-      ErrorVariant.http410,
-      Some(
-        """"
-          |This API is now decommissioned.
-          |Please use https://api.wellcomecollection.org/catalogue/v2/works.
-        """.stripMargin.replace('\n', ' ')
-      )
-    )
-  )
-
-  def notFound = extractPublicUri { uri =>
-    error(
-      DisplayError(
-        ErrorVariant.http404,
-        Some(s"Page not found for URL ${uri.path}")
-      )
+      HttpEntity(MediaTypes.`application/json`, swaggerDocs.json)
     )
   }
 
-  def invalidParam(msg: String) = error(
-    DisplayError(ErrorVariant.http400, Some(s"$msg"))
-  )
-
-  def elasticError(err: ElasticError): Route = error(
-    ElasticsearchErrorHandler.buildDisplayError(err)
-  )
-
-  def error(err: DisplayError): Route = {
-    val status = err.httpStatus match {
-      case Some(400) => BadRequest
-      case Some(404) => NotFound
-      case Some(410) => Gone
-      case Some(500) => InternalServerError
-      case _         => InternalServerError
-    }
-    complete(status -> ResultResponse(context = contextUri, result = err))
-  }
+  val swaggerDocs = new SwaggerDocs(apiConfig)
 
   def getClusterHealth: Route = {
     import com.sksamuel.elastic4s.ElasticDsl._
@@ -215,81 +86,16 @@ class Router(elasticClient: ElasticClient,
     }
   }
 
-  def swagger: Route = get {
-    complete(
-      HttpEntity(MediaTypes.`application/json`, swaggerDocs.json)
-    )
-  }
-
   def rejectionHandler =
     RejectionHandler.newBuilder
       .handle {
         case MalformedQueryParamRejection(field, msg, _) =>
-          invalidParam(s"$field: $msg")
+          invalidRequest(s"$field: $msg")
         case ValidationRejection(msg, _) =>
-          invalidParam(s"$msg")
+          invalidRequest(s"$msg")
       }
-      .handleNotFound(notFound)
+      .handleNotFound(extractPublicUri { uri =>
+        notFound(s"Page not found for URL ${uri.path}")
+      })
       .result
-
-  def getWithFuture(future: Future[Route]): Route =
-    get {
-      onComplete(future) {
-        case Success(resp) => resp
-        case Failure(err) =>
-          error(
-            DisplayError(ErrorVariant.http500, Some("Unhandled error."))
-          )
-      }
-    }
-
-  def retrieveTree(
-    index: Index,
-    work: IdentifiedWork,
-    expandedPaths: List[String]): Future[Option[(Collection, List[String])]] =
-    work.data.collectionPath
-      .map {
-        case CollectionPath(path, _, _) =>
-          val allPaths = path :: expandedPaths
-          collectionService.retrieveTree(index, allPaths).map {
-            case Left(err) =>
-              // We just log this here rather than raising so as not to bring down
-              // the work API when tree retrieval fails
-              logger.error("Error retrieving collection tree", err)
-              None
-            case Right(tree) => Some((tree, allPaths))
-          }
-      }
-      .getOrElse(Future.successful(None))
-
-  // Directive for getting public URI of the current request, using the host
-  // and scheme as per the config.
-  // (without this URIs end up looking like https://localhost:8888/..., rather
-  // than https://api.wellcomecollection.org/...))
-  def extractPublicUri =
-    extractUri.map { uri =>
-      uri
-        .withHost(apiConfig.host)
-        // akka-http uses 0 to indicate no explicit port in the URI
-        .withPort(0)
-        .withScheme(apiConfig.scheme)
-    }
-
-  val swaggerDocs = new SwaggerDocs(apiConfig)
-
-  lazy val contextUri =
-    apiConfig match {
-      case ApiConfig(host, scheme, _, pathPrefix, contextSuffix) =>
-        s"$scheme://$host/$pathPrefix/${ApiVersions.v2}/$contextSuffix"
-    }
-
-  lazy val worksService =
-    new WorksService(new ElasticsearchService(elasticClient))
-
-  lazy val collectionService =
-    new CollectionService(elasticClient)
-
-  lazy val logger = Logger(this.getClass.getName)
-
-  implicit val jsonPrinter: Printer = DisplayJsonUtil.printer
 }
